requirements.txt
langchain
langchain-google-genai
langchain-core
python-dotenv
pydantic